Post Observer

+++ routes/web.php
@@ -36,10 +36,9 @@ Route::post('/login', [SessionsController::class, 'store'])->middleware('guest')
 Route::post('/newsteller', NewsletterController::class);
 
 
-Route::get('admin/posts/create', [PostController::class, 'create'])->middleware('admin');
-Route::post('admin/posts/create', [PostController::class, 'store'])->middleware('admin');
-
-Route::get('admin/posts', [AdminPostController::class, 'index']);
-Route::get('admin/posts/{post}/edit', [AdminPostController::class, 'edit']);
-Route::patch('admin/posts/{post}/edit', [AdminPostController::class, 'update']);
-Route::delete('/admin/posts/{post}', [AdminPostController::class, 'destroy']);
+Route::get('admin/posts/create', [AdminPostController::class, 'create'])->middleware('admin');
+Route::post('admin/posts/create', [AdminPostController::class, 'store'])->middleware('admin');
+Route::get('admin/posts', [AdminPostController::class, 'index'])->middleware('admin');
+Route::get('admin/posts/{post}/edit', [AdminPostController::class, 'edit'])->middleware('admin');
+Route::patch('admin/posts/{post}/edit', [AdminPostController::class, 'update'])->middleware('admin');
+Route::delete('/admin/posts/{post}', [AdminPostController::class, 'destroy'])->middleware('admin');
